Added configuration of update site, fixed no proxy host configuration
